Feat:공간 기록 마무리 단계에서 큐레이션에 장소 추가될 수 있도록 기능 구현
